+++ packages/datadog-plugin-cypress/src/after-run.js
@@ -1,5 +1,3 @@
 const { cypressPlugin } = require('./cypress-plugin')
 
-module.exports = {
-  afterRun: cypressPlugin.afterRun.bind(cypressPlugin)
-}
+module.exports = cypressPlugin.afterRun.bind(cypressPlugin)
